requirements.txt
alembic==1.6.5
astroid==2.5.6
bleach==6.1.0
blinker==1.6.2
click==8.1.6
dataclasses==0.6
dnspython==2.6.1
email-validator==1.1.2
Flask==2.3.2
Flask-Login==0.6.3
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
greenlet==3.0.0
gunicorn==22.0.0
idna==3.7
importlib-metadata==4.11.3
isort==5.8.0
itsdangerous==2.1.2
Jinja2==3.1.6
lazy-object-proxy==1.6.0
Mako==1.2.4
Markdown==3.3.6
MarkupSafe==2.1.3
mccabe==0.6.1
mysqlclient==2.1.1
pip==25.3
pylint==2.8.3
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.6
PyMySQL==1.1.1
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
q==2.6
setuptools==78.1.1
six==1.16.0
SQLAlchemy==1.4.17
toml==0.10.2
typing_extensions==4.5.0
Werkzeug==3.0.6
wrapt==1.12.1
WTForms==2.3.3
zipp==3.19.1
